Fixed KMP being in wrong directory
